test(simulation): split browser contracts by responsibility
Move the existing eight Analog Simulation browser cases into plot export, setup lifecycle, batch/run-plan, human workspace, and Agent raw suites. Centralize the read-only OTA/Profile fixtures and route Analog and Agent changes without locally running the same spec twice. Test bodies are byte-equivalent after line-ending normalization; no scenario or assertion was added or removed.

Validation: 8 Analog Simulation Playwright cases passed

Validation: validation planner contracts passed (26)

Validation: mechanical test-body comparison old=8 current=8 identical=true

Test-Impact: tests-updated
