+++ apps/editor/e2e/simulation-e2e-fixtures.ts
@@ -1,0 +1,26 @@
+import { readFileSync } from "node:fs";
+
+export const profile = JSON.parse(
+  readFileSync(
+    new URL(
+      "../../../containers/ngspice/hosted-sky130-profile.json",
+      import.meta.url,
+    ),
+    "utf8",
+  ),
+) as {
+  id: string;
+  displayName: string;
+  simulator: { version: string };
+  models: { library: { runtimePath: string } };
+};
+
+export const ota = JSON.parse(
+  readFileSync(
+    new URL(
+      "../src/examples/five-transistor-ota-sky130.icproj.json",
+      import.meta.url,
+    ),
+    "utf8",
+  ),
+);
